requirements.txt
numpy
pandas
python-dotenv
requests
librosa